pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "qadmcli"
version = "0.1.0"
description = "AS400 DB2 for i Database Management CLI Tool"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "QADM CLI Team" }
]
keywords = ["as400", "db2", "iseries", "ibm-i", "database", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "jaydebeapi>=1.2.3",
    "jpype1>=1.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
qadmcli = "qadmcli.cli:main"

[project.urls]
Homepage = "https://github.com/qoder/qadmcli"
Repository = "https://github.com/qoder/qadmcli"
Issues = "https://github.com/qoder/qadmcli/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/qadmcli"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "W", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
